Move to munit in query module
