+++ project/Dependencies.scala
@@ -28,6 +28,7 @@ object Dependencies {
   val LogbackVersion          = "1.2.3"
   val MariaDbVersion          = "2.7.2"
   val MiniTestVersion         = "2.9.3"
+  val MUnitVersion            = "0.7.22"
   val OrganizeImportsVersion  = "0.5.0"
   val PdfboxVersion           = "2.0.22"
   val PoiVersion              = "4.1.2"
@@ -270,6 +271,11 @@ object Dependencies {
     "io.monix" %% "minitest"      % MiniTestVersion,
     "io.monix" %% "minitest-laws" % MiniTestVersion
   ).map(_ % Test)
+
+  val munit = Seq(
+    "org.scalameta" %% "munit" % MUnitVersion,
+    "org.scalameta" %% "munit-scalacheck" % MUnitVersion
+  )
 
   val kindProjectorPlugin = "org.typelevel" %% "kind-projector"     % KindProjectorVersion
   val betterMonadicFor    = "com.olegpy"    %% "better-monadic-for" % BetterMonadicForVersion
